further USB cleanup

+++ src/natfeat/usbhost.h
@@ -157,7 +157,7 @@ class USBHost : public NF_Base
 {
 private:
 	int32 rh_devnum;			/* address of Root Hub endpoint */
-	int8 total_num_handles;
+	int total_num_handles;
 
 	
 	int32 aranym_submit_rh_msg(usb_device *dev, uint32 pipe,
